requirements.txt
streamlit
openai-whisper
moviepy>=2.0.0
pydub
gtts
deep-translator
ffmpeg-python
torch
